refactor: RepositoryUiModel github.model 패키지로 이동
